# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Club deportivo API",
        "description": "API Reservas club deportivo, es  una pureba tecnica para la empresa NETSNITS ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Reservas"
        },
        {
            "url": "http://localhost:8000",
            "description": "Class DeportesController"
        },
        {
            "url": "http://localhost:8000",
            "description": "Class PistasController"
        },
        {
            "url": "http://localhost:8000",
            "description": "Class ReservasController"
        },
        {
            "url": "http://localhost:8000",
            "description": "Class SociosController"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/buscador": {
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "buscar reservas",
                "description": "Muestra una busqueda de reservas basada en la fecha especificada ex 2023-11-20",
                "operationId": "60da9e9dc952910e1947810100da91da",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "query",
                        "description": "Fecha en formato 'Y-m-d' (ejemplo: '2023-01-01')",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reservas": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "reserva": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "socio": {
                                                                    "type": "string"
                                                                },
                                                                "pista": {
                                                                    "type": "string"
                                                                },
                                                                "deporte": {
                                                                    "type": "string"
                                                                },
                                                                "fecha": {
                                                                    "type": "string",
                                                                    "format": "date"
                                                                },
                                                                "horaInicio": {
                                                                    "type": "string",
                                                                    "format": "time"
                                                                },
                                                                "horaFin": {
                                                                    "type": "string",
                                                                    "format": "time"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El campo fecha debe tener el formato 'Y-m-d'."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error interno"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}